requirements.txt
flask==3.0.0
flask-cors==4.0.0
Flask-HTTPAuth==4.8.0
openai>=1.55.0
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
Pillow>=10.0.0


